requirements.txt
accelerate==1.1.1
aiohttp==3.10.5
aiohttp==3.8.3
aiohttp==3.8.5
aiohttp==3.9.0
apex==0.9.10dev
attrs==22.1.0
audiocraft==1.3.0
av==13.1.0
azureml==0.2.7
beautifulsoup4==4.12.3
Brotli==1.0.9
brotlicffi==1.1.0.0
brotlipy==0.7.0
causal_conv1d==1.4.0
classy_vision==0.7.0
clearml==1.16.5
codecarbon==2.8.0
comet_ml==3.47.3
compressed_tensors==0.8.0
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==41.0.2
cryptography==41.0.3
cryptography==41.0.7
ctypes_snappy==1.03
Cython==3.0.11
dagshub==0.3.45
dall_e==0.1
distributed==2023.11.0
distributed==2024.8.2
distributed==2023.6.0
Django==4.1
Django==5.0.8
Django==5.1.1
django_recaptcha==4.0.0
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
duckduckgo_search==6.3.7
dvclive==3.48.0
einops==0.8.0
essentia==2.1b6.dev1177
fairseq==0.12.2
fairseq2==0.2.1
faiss==1.5.3
fastai==2.7.18
fastapi==0.115.5
fastparquet==2024.11.0
fbgemm_gpu==1.0.0
flash_attn==2.7.0.post2
flatdict==4.0.1
flytekit==1.13.14
ftfy==6.3.1
fugashi==1.4.0
g2p_en==2.1.0
galore_torch==1.0
gdown==5.2.0
gensim==4.3.3
gensim==4.3.0
gguf==0.10.0
gluonnlp==0.10.0
gradio==5.6.0
grokadamw==0.1.2
h2==4.1.0
h5py==3.7.0
h5py==3.9.0
h5py==3.11.0
haiku==0.01
hf_transfer==0.1.8
HTMLParser==0.0.2
hypothesis==6.119.4
InquirerPy==0.3.4
intel_extension_for_pytorch==2.5.0
ipadic==1.0.0
ipython==8.12.0
ipython==8.15.0
ipython==8.20.0
ipython==8.27.0
ipython==8.12.3
ipywidgets==8.1.2
ipywidgets==8.0.4
isal==1.7.1
jax==0.4.35
jaxlib==0.4.35
jieba==0.42.1
Jinja2==3.1.2
jnius==1.1.0
js==1.0
kenlm==0.2.0
keras_nlp==0.17.0
keras_tuner==1.4.7
kerchunk==0.2.7
keyring==24.3.1
keyring==23.13.1
laion_clap==1.1.6
librosa==0.10.2.post1
liger_kernel==0.4.2
lockfile==0.12.2
lomo_optim==0.1.1
lz4==4.3.2
lzmaffi==0.3.0
mamba_ssm==2.2.2
mambapy==1.2.0
markdownify==0.14.1
Mask2Former==0.1.5
matplotlib==3.7.1
matplotlib==3.7.2
matplotlib==3.8.0
megatron==0.5.1
mistral_common==1.5.1
mlflow==2.18.0
mock==5.1.0
mtrand==0.1
MultiScaleDeformableAttention==1.0
mxnet==1.9.1
mypy==1.13.0
natten==0.17.3
nemo==6.0.3
neptune==1.13.0
nltk==3.9.1
nltk==3.8.1
numarray==1.5.1
Numeric==24.2
onnxruntime==1.20.1
optimum==1.23.3
optuna==4.1.0
pandas==2.0.3
pandas==1.5.3
paramiko==2.8.1
phonemizer==3.3.0
pickle5==0.0.12
Pillow==9.4.0
Pillow==10.0.1
Pillow==11.0.0
pretty_midi==0.2.10
protobuf==5.28.3
psutil==5.9.0
py3nvml==0.2.7
pyarrow==14.0.2
pyarrow==11.0.0
pycocotools==2.0.8
pyctcdecode==0.5.0
pydantic==1.10.12
pygit2==1.16.0
pyinstaller==6.6.0
pyodide==0.0.2
pyOpenSSL==24.2.1
pyOpenSSL==23.2.0
pyOpenSSL==24.2.1
pytesseract==0.3.13
pytest==7.4.4
pytest==7.4.0
pytest.egg==info
pythainlp==5.0.4
python_Levenshtein==0.26.1
pytorch_lightning==2.4.0
pytorch_quantization==2.2.1
pytz==2022.7
pytz==2024.1
pytz==2023.3.post1
pywin32==305.1
railroad==0.5.0
ray==2.39.0
redis==5.2.0
requests_kerberos==0.15.0
rhoknp==1.7.0
rjieba==0.1.11
s3prl==0.4.16
sacremoses==0.0.43
scann==1.3.4
schedulefree==1.3
scikit_learn==1.2.2
scikit_learn==1.3.0
scipy==1.11.1
scipy==1.11.4
scipy==1.10.1
scipy_doctest==1.4
sentencepiece==0.2.0
setuptools_scm==8.1.0
sigopt==8.8.3
simplejson==3.19.3
slack_sdk==3.33.4
smbprotocol==1.15.0
soundfile==0.12.1
spacy==3.8.2
Sphinx==5.0.2
starlette==0.41.3
sudachipy==0.6.9
sympy==1.11.1
sympy==1.13.2
sympy==1.12
t5x==0.0.0
tensorboardX==2.6.2.2
tensorflow_hub==0.16.1
tensorflow_probability==0.25.0
tensorflow_text==2.18.0
tensorstore==0.1.69
tf2onnx==1.16.1
tf_keras==2.18.0
thread==2.0.5
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tokenizer==3.4.5
toml==0.10.2
torch_npu==2.4.0
torch_xla==2.5.1
torchao==0.6.1
torchaudio==2.0.2
torchaudio==2.3.1
torchaudio==2.3.1+cu121
torchvision==0.15.2
torchvision==0.18.1
torchvision==0.18.1+cu121
tornado==6.3.2
tornado==6.4.1
tornado==6.3.3
trajectory==0.1.1
trove_classifiers==2024.10.21.16
ujson==5.10.0
ujson==5.4.0
unidic==1.1.0
unidic_lite==1.0.8
urllib3_secure_extra==0.1.0
uroman==1.3.1.1
uvicorn==0.32.1
vissl==0.1.6
wandb==0.18.7
wget==3.2
xformers==0.0.28.post3
xmlrpclib==1.0.1
yarl==1.11.0
yarl==1.8.1
yarl==1.9.3
zstandard==0.22.0
zstandard==0.19.0
